refactor: split flow cache types from types.rs into flow_cache.rs
Move FlowCache, FlowCacheEntry, RewriteDescriptor + 10 tests into
afxdp/flow_cache.rs (716 lines). types.rs drops from 1302 to 586.

Closes #367
